Make deployment jobs compatible with standalone and integrated mode
We are working in integrating the watcher-operator in the
openstack-operator following the same approach that other openstack
services operators are doing.

To make the transition smooth, I am proposing to make the jobs
compatible with both standalone and integrated mode by implementing
following changes:

- Now, the content provider jobs should be rebuilding the
  watcher-operator and the openstack-operator. For the integration mode,
  that should make that, when the openstack-operator deployes the
  watcher-operator it will run using the patched version of it.
- I'm adding logic to find out if the watcher-operator is already
  deployed when running the deploy_watcher_service.yaml playbook. If the
  Watcher CRDs are already installed, it means the watcher operator has
  been installed by the openstack-operator and there is no need to install
  it standalone.
- I'm also adding logic to find out if the watcher is integrated in the
  openstackcontrolplane. If so, watcher is deployed by patching the
  existing openstackcontrolplane. Otherwise, it creates a Watcher CR
  as it is currently doing.
- Finally, we need to cover the cases where we deploy Watcher using a
  Watcher CR but there are no matating webhooks to include the container
  images automatically (that will be managed at the
  openstackcontrolplane level when fully integrated).

Note that, at some point we will be able to improve this approach by
applying a kustomization to the controlplane in pre_deploy, but at
the pre_deploy is not possible to discover if the watcher is already
integrated in the openstackcontrolplane CRD, so we can keep using this
post_deployment approach until the integration is finished and
propagated to all the branches and environments.

Jira: OSPRH-16242